+++ src/main/native-chat/transcript-line-decoders-claude.ts
@@ -1,0 +1,55 @@
+// Claude JSONL line → NativeChatMessage decoder.
+
+import type { NativeChatBlock, NativeChatMessage } from '../../shared/native-chat-types'
+import {
+  asRecord,
+  extractString,
+  parseJsonObject,
+  timestampMs
+} from '../ai-vault/session-scanner-values'
+import { claudeContentBlocks } from './transcript-record-blocks'
+
+export function decodeClaudeTranscriptLine(
+  line: string,
+  fallbackId: string
+): NativeChatMessage | null {
+  const record = parseJsonObject(line)
+  if (!record) {
+    return null
+  }
+  const role = record.type
+  if (role !== 'user' && role !== 'assistant') {
+    return null
+  }
+  const message = asRecord(record.message)
+  const blocks = claudeContentBlocks(message?.content)
+  if (blocks.length === 0) {
+    return null
+  }
+  const messageId = extractString(record.uuid) ?? extractString(message?.id)
+  return {
+    id: messageId ?? fallbackId,
+    role: claudeMessageRole(role, blocks),
+    blocks,
+    timestamp: parseTimestamp(record.timestamp),
+    source: 'transcript'
+  }
+}
+
+// Claude marks reasoning via `thinking` content blocks; when a message is made
+// up solely of reasoning, surface it as a reasoning-role message.
+function claudeMessageRole(
+  role: 'user' | 'assistant',
+  blocks: NativeChatBlock[]
+): NativeChatMessage['role'] {
+  if (role === 'user') {
+    const onlyToolResults = blocks.every((block) => block.type === 'tool-result')
+    return onlyToolResults && blocks.length > 0 ? 'tool' : 'user'
+  }
+  return role
+}
+
+function parseTimestamp(value: unknown): number | null {
+  const parsed = timestampMs(value)
+  return Number.isFinite(parsed) ? parsed : null
+}
